Rename the controller package to engine, rename types to not stutter
